chore: migrate remaining tests from tests-ui/tests to src/
Amp-Thread-ID: https://ampcode.com/threads/T-019b7267-bf86-7241-a26e-4fe29fc9b4c6
