+++ src/extensions/core/dynamicPrompts.test.ts
