+++ src/scripts/api.folderPaths.test.ts
