+++ src/stores/domWidgetStore.test.ts
